+++ src/main/java/com/li64/tide/registries/items/EnchantedPocketWatchItem.java
@@ -2,6 +2,7 @@ package com.li64.tide.registries.items;
 
 import com.li64.tide.data.FreezableMob;
 import com.li64.tide.data.TideCriteriaTriggers;
+import com.li64.tide.data.TideTags;
 import com.li64.tide.registries.TideParticleTypes;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.registries.BuiltInRegistries;
@@ -28,7 +29,11 @@ public class EnchantedPocketWatchItem extends PocketWatchItem {
 
     @Override
     public @NotNull InteractionResult interactLivingEntity(ItemStack stack, Player player, LivingEntity target, InteractionHand hand) {
-        if (!(target instanceof Mob mob) || !(target instanceof FreezableMob freezable) || player.getCooldowns().isOnCooldown(this)) return super.interactLivingEntity(stack, player, target, hand);
+        if (!(target instanceof Mob mob) || !(target instanceof FreezableMob freezable)
+                || target.getType().is(TideTags.Entities.IGNORES_POCKET_WATCH)
+                || player.getCooldowns().isOnCooldown(this)) {
+            return super.interactLivingEntity(stack, player, target, hand);
+        }
         if (!(player.level() instanceof ServerLevel level)) return InteractionResult.SUCCESS;
         if (freezable.tide$isFrozen()) {
             level.playSound(null, mob.blockPosition(), SoundEvents.CHAIN_BREAK, SoundSource.PLAYERS, 1.0f, 1.0f);